Add collection of shm allocators with various strong sides
