+++ Cargo.toml
@@ -90,6 +90,7 @@ async-trait = "0.1.82"
 base64 = "0.22.1"
 bincode = "1.3.3"
 bytes = "1.7.1"
+buddy_system_allocator = "0.11.0"
 clap = { version = "4.5.17", features = ["derive"] }
 console-subscriber = "0.4.0"
 const_format = "0.2.33"
@@ -170,6 +171,7 @@ static_assertions = "1.1.0"
 static_init = "1.0.3"
 stop-token = "0.7.0"
 syn = "2.0"
+talc = { version = "4.4.3", default-features = false }
 test-case = "3.3.1"
 tide = "0.16.0"
 time = "0.3.36"
